Create restaurantsDAO.js
Create a directory "dao" which is a data access object inside the api folder. And then, create the file  "restaurantsDAO.js" inside the dao folder.

The variable restaurant is used to store references in our database. "injectDB" method is used to initially connect to our database and we have gone a call this method as restaurants.server strat. So, when restaurants.server start we going to get the reference to our restaurant databases. And then the other function the "getRestaurants" get the list of restaurants in the database with some filters.