Update restaurants.route.js
Add two more routes to the route file for the restaurants. here, route the specific restaurant to get all the reviews associated with that restaurant. And, router cuisines to return a list of cuisines. The reason is in the frontend we want a user to be a select the cuisines from a dropdown menu indeed of having all the cuisines memorized.