Update add-review.js
initial review state set as empty. Then check the new review or the editing review.  If we are in an editing review edit the review.

Set the submitted ad false in the initial state. Then handle the Input for review.

There is a form to edit or create the review.

Save review function used to update data.